Refactor UniFi Network Rules switch module into multiple files for improved organization
- Decomposed the existing switch.py into dedicated modules for various switch types (firewall, traffic route, static route, etc.).
- Introduced a new switches package to maintain backward compatibility with existing functionality.
- Updated imports to ensure seamless integration with the refactored structure.
- Enhanced code clarity and maintainability by separating concerns across different files.
