+++ custom_components/unifi_network_rules/switches/firewall.py
@@ -1,0 +1,39 @@
+"""Firewall switches for UniFi Network Rules integration."""
+from __future__ import annotations
+
+from typing import Any
+
+from .base import UnifiRuleSwitch
+from ..coordinator import UnifiRuleUpdateCoordinator
+
+
+class UnifiFirewallPolicySwitch(UnifiRuleSwitch):
+    """Switch to enable/disable a UniFi firewall policy."""
+    
+    def __init__(
+        self,
+        coordinator: UnifiRuleUpdateCoordinator,
+        rule_data: Any,
+        rule_type: str,
+        entry_id: str = None,
+    ) -> None:
+        """Initialize firewall policy switch."""
+        super().__init__(coordinator, rule_data, rule_type, entry_id)
+        # Set icon for firewall policy rules
+        self._attr_icon = "mdi:shield-check"
+
+
+class UnifiLegacyFirewallRuleSwitch(UnifiRuleSwitch):
+    """Switch to enable/disable a UniFi legacy firewall rule."""
+    
+    def __init__(
+        self,
+        coordinator: UnifiRuleUpdateCoordinator,
+        rule_data: Any,
+        rule_type: str,
+        entry_id: str = None,
+    ) -> None:
+        """Initialize legacy firewall rule switch."""
+        super().__init__(coordinator, rule_data, rule_type, entry_id)
+        # Set icon for legacy firewall rules
+        self._attr_icon = "mdi:shield-outline"
